refactor: SessionQueryClient
